+++ lib/models/danmaku_option.dart
@@ -9,7 +9,7 @@ class DanmakuOption {
   final double area;
 
   /// 滚动弹幕运行时间，秒
-  final int duration;
+  final double duration;
 
   /// 不透明度，0.1-1.0
   final double opacity;
@@ -53,7 +53,7 @@ class DanmakuOption {
     double? fontSize,
     int? fontWeight,
     double? area,
-    int? duration,
+    double? duration,
     double? opacity,
     bool? hideTop,
     bool? hideBottom,
